chore(deps): bump object_store from 0.11 to 0.14 (#308)
* chore(deps): bump object_store from 0.11 to 0.13

Supersedes #237. object_store 0.12/0.13 moved several convenience
methods (get, head, get_range, put_multipart, delete, copy,
copy_if_not_exists) off the ObjectStore trait and onto the new
ObjectStoreExt trait, and switched offset/range accounting from
usize to u64. copy/copy_if_not_exists collapsed into copy_opts.

Production call sites just import ObjectStoreExt where needed and
cast the resume offset to u64. The two test-only ObjectStore impls
lose the delegating overrides for moved-out methods, replace
copy/copy_if_not_exists with a single copy_opts, gain a delete_stream
delegate, and switch the seen_offsets vec to u64.

Also picks up new quick-xml versions that resolve the two RUSTSEC
advisories flagged against 0.37.5.

* chore(deps): bump object_store to 0.14, install rustls provider

Push through to 0.14.0 rather than 0.13.2. 0.14 added an `extensions:
Extensions` field to `GetResult` — set to `Default::default()` in
the ResumeStore test impl to match.

The 0.14 dep chain also bumps reqwest 0.12 → 0.13, which brings in a
rustls with both `ring` and `aws-lc-rs` compiled in via transitive
features. rustls' `default_provider()` auto-select now panics with
`Could not automatically determine the process-level CryptoProvider`
unless the process has explicitly picked one. NetService::new installs
the ring provider once via Once — same choice we already make in
tls::PeerVerifier, just made visible to code paths that go through
rustls' process-global lookup.

* net/svc,ci: install rustls provider from PeerVerifier + ignore quick-xml advisories

The install hook needs to fire on the direct TLS-config path too, not
just NetService::new — tls::tests::make_configs_works builds a
config directly and hit the same panic. Move it into PeerVerifier::new
(every server/client config path goes through it) and drop the
duplicate call in NetService::new.

Also add the three transitively-vulnerable quick-xml advisories to
.cargo/audit.toml. The fix is in quick-xml >= 0.41.0 and object_store
0.14 still pins 0.40.x — until upstream bumps, ignoring is the
pragmatic option. Justification is in-file: mosaic only feeds trusted
S3 responses into object_store, so the DoS surface is not reachable
via attacker-controlled XML.

+++ crates/storage/s3/src/reader.rs
@@ -12,7 +12,7 @@ use std::{future::Future, sync::Arc};
 use futures::StreamExt;
 use mosaic_common::Byte32;
 use mosaic_storage_api::table_store::{TableMetadata, TableReader};
-use object_store::{GetOptions, GetRange, ObjectStore};
+use object_store::{GetOptions, GetRange, ObjectStore, ObjectStoreExt};
 use tracing::{debug, error, warn};
 
 use crate::{error::S3Error, paths::TableRootPaths};
@@ -283,7 +283,7 @@ async fn stream_ciphertexts(
             GetOptions::default()
         } else {
             GetOptions {
-                range: Some(GetRange::Offset(bytes_delivered)),
+                range: Some(GetRange::Offset(bytes_delivered as u64)),
                 ..Default::default()
             }
         };
